requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
appdirs==1.4.4
asttokens==2.4.1
attrs==24.2.0
bio==1.7.1
biopython==1.84
biothings-client==0.3.1
blinker==1.8.2
cachetools==5.5.0
CAI-PyPI==2.0.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
CodonTransformer==1.2.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
forgi==2.2.3
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gprofiler-official==1.0.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lightning-utilities==0.11.7
logging_exceptions==0.1.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mygene==3.2.2
narwhals==1.6.3
networkx==3.3
numpy==2.1.1
onnxruntime==1.19.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==5.28.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-codon-tables==0.1.12
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
torchmetrics==1.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
ViennaRNA==2.6.4
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
widgetsnbextension==4.0.13
yarl==1.11.0
